// src/graphql/generated/schema.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string | number; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};
// ...
export type Company = {
  __typename?: 'Company';
  avater?: Maybe<Scalars['String']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  employee?: Maybe<Array<Maybe<Employee>>>;
  id?: Maybe<Scalars['String']['output']>;
  logo?: Maybe<Scalars['String']['output']>;
  name?: Maybe<Scalars['String']['output']>;
  owner?: Maybe<User>;
  ownerId?: Maybe<Scalars['String']['output']>;
  slug?: Maybe<Scalars['String']['output']>;
};

export type CompanyFilterInput = {
  search?: InputMaybe<Scalars['String']['input']>;
};

export type CompanyInput = {
  avater: Scalars['String']['input'];
  description: Scalars['String']['input'];
  logo: Scalars['String']['input'];
  name: Scalars['String']['input'];
  slug: Scalars['String']['input'];
};

export enum CompanyRole {
owner ,employee
}

export type CompanyUser = {
  __typename?: 'CompanyUser';
  companyRole?: Maybe<CompanyRole>;
  email?: Maybe<Scalars['String']['output']>;
  firstname?: Maybe<Scalars['String']['output']>;
  id?: Maybe<Scalars['String']['output']>;
  lastname?: Maybe<Scalars['String']['output']>;
  role?: Maybe<UserRole>;
};
// ...
export type GeoLocationInput = {
  companyId?: InputMaybe<Scalars['String']['input']>;
  latitude?: InputMaybe<Scalars['Float']['input']>;
  longitude?: InputMaybe<Scalars['Float']['input']>;
};

export type Location = {
  __typename?: 'Location';
  city?: Maybe<Scalars['String']['output']>;
  companyId?: Maybe<Scalars['String']['output']>;
  country?: Maybe<Scalars['String']['output']>;
  id?: Maybe<Scalars['String']['output']>;
  state?: Maybe<Scalars['String']['output']>;
  streetAddress?: Maybe<Scalars['String']['output']>;
  zip?: Maybe<Scalars['String']['output']>;
};

export type LocationInput = {
  city?: InputMaybe<Scalars['String']['input']>;
  country?: InputMaybe<Scalars['String']['input']>;
  state?: InputMaybe<Scalars['String']['input']>;
  streetAddress?: InputMaybe<Scalars['String']['input']>;
  zip?: InputMaybe<Scalars['String']['input']>;
};

export type LoginInput = {
  email?: InputMaybe<Scalars['String']['input']>;
  password?: InputMaybe<Scalars['String']['input']>;
};

export type Mutation = {
  __typename?: 'Mutation';
  addPaymentMethod?: Maybe<Scalars['String']['output']>;
  createBooking?: Maybe<DefaultResposnce>;
  createCompany?: Maybe<DefaultResposnce>;
  createCompanyWithOwner?: Maybe<CompanyLoginResponsce>;
  createEmployee?: Maybe<DefaultResposnce>;
  createGeoLocation?: Maybe<DefaultResposnce>;
  createLocation?: Maybe<DefaultResposnce>;
  createPlan?: Maybe<DefaultResposnce>;
  createService?: Maybe<DefaultResposnce>;
  createsubscription?: Maybe<DefaultResposnce>;
  deleteBooking?: Maybe<DefaultResposnce>;
  deleteService?: Maybe<DefaultResposnce>;
  editCompany?: Maybe<DefaultResposnce>;
  editPlan?: Maybe<DefaultResposnce>;
  login?: Maybe<Loginresponsce>;
  ownerLogin?: Maybe<Loginresponsce>;
  ownerRegister?: Maybe<DefaultResposnce>;
  register?: Maybe<DefaultResposnce>;
  updateBooking?: Maybe<DefaultResposnce>;
  updateService?: Maybe<DefaultResposnce>;
  updatesubscription?: Maybe<DefaultResposnce>;
};
// ...
export type OwnerInput = {
  name?: InputMaybe<Scalars['String']['input']>;
};
// ...
export type User = {
  __typename?: 'User';
  email?: Maybe<Scalars['String']['output']>;
  firstname?: Maybe<Scalars['String']['output']>;
  id?: Maybe<Scalars['String']['output']>;
  lastname?: Maybe<Scalars['String']['output']>;
  role?: Maybe<UserRole>;
};

export type UserInput = {
  confirmPassword?: InputMaybe<Scalars['String']['input']>;
  email?: InputMaybe<Scalars['String']['input']>;
  firstname?: InputMaybe<Scalars['String']['input']>;
  lastname?: InputMaybe<Scalars['String']['input']>;
  password?: InputMaybe<Scalars['String']['input']>;
};

export enum UserRole {
  Admin = 'admin',
  Company = 'company',
  Public = 'public',
  Superadmin = 'superadmin'
}

export type CompanyLoginResponsce = {
  __typename?: 'companyLoginResponsce';
  accessToken?: Maybe<Scalars['String']['output']>;
  isAuthenticated?: Maybe<Scalars['Boolean']['output']>;
  message?: Maybe<Scalars['String']['output']>;
  success?: Maybe<Scalars['Boolean']['output']>;
  user?: Maybe<CompanyUser>;
};

export enum CompanyRole {
  Employee = 'employee',
  Owner = 'owner'
}

export type DefaultResposnce = {
  __typename?: 'defaultResposnce';
  message?: Maybe<Scalars['String']['output']>;
  success?: Maybe<Scalars['Boolean']['output']>;
};
// ...
export type Loginresponsce = {
  __typename?: 'loginresponsce';
  accessToken?: Maybe<Scalars['String']['output']>;
  isAuthenticated?: Maybe<Scalars['Boolean']['output']>;
  message?: Maybe<Scalars['String']['output']>;
  success?: Maybe<Scalars['Boolean']['output']>;
  user?: Maybe<User>;
};
// ...
export type LoginMutation = { __typename?: 'Mutation', login?: { __typename?: 'loginresponsce', accessToken?: string | null, message?: string | null, isAuthenticated?: boolean | null, success?: boolean | null, user?: { __typename?: 'User', id?: string | null, firstname?: string | null, lastname?: string | null, email?: string | null, role?: UserRole | null } | null } | null };

export type CreateCompanyWithOwnerMutationVariables = Exact<{
  userInput?: InputMaybe<UserInput>;
  ownerInput?: InputMaybe<OwnerInput>;
  companyInput?: InputMaybe<CompanyInput>;
  locationInput?: InputMaybe<LocationInput>;
  geoLocationInput?: InputMaybe<GeoLocationInput>;
}>;


export type CreateCompanyWithOwnerMutation = { __typename?: 'Mutation', createCompanyWithOwner?: { __typename?: 'companyLoginResponsce', accessToken?: string | null, isAuthenticated?: boolean | null, message?: string | null, success?: boolean | null, user?: { __typename?: 'CompanyUser', email?: string | null, firstname?: string | null, id?: string | null, lastname?: string | null, role?: UserRole | null, companyRole?: CompanyRole | null } | null } | null };
// ...
export const LoginDocument = gql`
    mutation Login($user: LoginInput) {
  login(user: $user) {
    accessToken
    message
    user {
      id
      firstname
      lastname
      email
      role
    }
    isAuthenticated
    success
  }
}
    `;
// ...
export const CreateCompanyWithOwnerDocument = gql`
    mutation CreateCompanyWithOwner($userInput: UserInput, $ownerInput: OwnerInput, $companyInput: CompanyInput, $locationInput: LocationInput, $geoLocationInput: GeoLocationInput) {
  createCompanyWithOwner(
    userInput: $userInput
    ownerInput: $ownerInput
    companyInput: $companyInput
    locationInput: $locationInput
    geoLocationInput: $geoLocationInput
  ) {
    accessToken
    isAuthenticated
    message
    success
    user {
      email
      firstname
      id
      lastname
      role
      companyRole
    }
  }
}
    `;